+++ src/main/java/org/crm/crmproject/controller/BoardController.java
@@ -3,7 +3,7 @@ package org.crm.crmproject.controller;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.crm.crmproject.dto.BoardDTO;
-import org.crm.crmproject.service.BoardService;
+import org.crm.crmproject.service.board.BoardService;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
