+++ src/main/java/org/crm/crmproject/domain/Review.java
@@ -16,7 +16,7 @@ public class Review {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long reviewId;
+    private Long reviewNo;
 
     //  ManyToOne 어노테이션으로 Customer, Store 양쪽 모두를 참조하게 설정
     @ManyToOne(fetch = FetchType.LAZY)
@@ -31,6 +31,20 @@ public class Review {
     @Column(nullable = false)
     @Min(1)
     @Max(5)
-    private int grade;
+    private int grade;  //  별점
+
+    //  리뷰 내용, 별점을 수정하는 기능
+    public void changeReview(String reviewText, int grade) {
+        this.reviewText = reviewText;
+        this.grade = grade;
+    }
+
+    public void setCustomer(Long customerNo) {
+        this.customer = Customer.builder().customerNo(customerNo).build();
+    }
+
+    public void setStore(Long storeNo) {
+        this.store = Store.builder().storeId(storeNo).build();
+    }
 
 }
