--- src/main/java/org/crm/crmproject/service/ReviewService.java
@@ -1,4 +1,0 @@
-package org.crm.crmproject.service;
-
-public interface ReviewService {
-}
