--- src/main/java/org/crm/crmproject/service/ReviewServiceImpl.java
@@ -1,11 +1,0 @@
-package org.crm.crmproject.service;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.log4j.Log4j2;
-import org.springframework.stereotype.Service;
-
-@Log4j2
-@Service
-@RequiredArgsConstructor
-public class ReviewServiceImpl implements ReviewService{
-}
