+++ src/main/java/org/crm/crmproject/service/menu/MenuService.java
@@ -1,0 +1,4 @@
+package org.crm.crmproject.service.menu;
+
+public interface MenuService {
+}
