+++ src/main/java/org/crm/crmproject/service/menu/MenuServiceImpl.java
@@ -1,11 +1,12 @@
-package org.crm.crmproject.service;
+package org.crm.crmproject.service.menu;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
+import org.crm.crmproject.service.SalesService;
 import org.springframework.stereotype.Service;
 
 @RequiredArgsConstructor
 @Log4j2
 @Service
-public class MenuServiceImpl implements SalesService{
+public class MenuServiceImpl implements SalesService {
 }
