+++ src/main/java/org/crm/crmproject/service/review/ReviewService.java
@@ -1,0 +1,12 @@
+package org.crm.crmproject.service.review;
+
+import org.crm.crmproject.domain.Customer;
+import org.crm.crmproject.domain.Store;
+import org.crm.crmproject.dto.ReviewDTO;
+
+import java.util.List;
+
+public interface ReviewService {
+
+
+}
