+++ src/main/resources/templates/reviews.html
@@ -1,0 +1,121 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>리뷰 페이지</title>
+    <style>
+        .stars {
+            display: inline-block;
+            width: 150px;
+            height: 30px;
+            background-image: url('https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.4/css/all.min.css');
+            background-size: contain;
+        }
+
+        .stars input {
+            display: none;
+        }
+
+        .stars label {
+            float: right;
+            cursor: pointer;
+            padding: 5px;
+            font-size: 24px;
+            color: #ccc;
+        }
+
+        .stars input:checked ~ label:before {
+            color: #ffc107;
+        }
+    </style>
+</head>
+<body>
+    <h1>Review Page</h1>
+
+    <h2>Add Review</h2>
+    <form id="addReviewForm">
+        <label for="customerId">Customer ID: </label>
+        <input type="text" id="customerId" name="customerId"><br><br>
+
+        <label for="storeId">Store ID:</label>
+        <input type="text" id="storeId" name="storeId"><br><br>
+
+        <label for="reviewText">Review Text:</label><br>
+        <textarea id="reviewText" name="reviewText"></textarea><br><br>
+
+        <div class="stars">
+            <input type="radio" id="star5" name="grade" value="5">
+            <label for="star5"><i class="fas fa-star"></i></label>
+            <input type="radio" id="star4" name="grade" value="5">
+            <label for="star4"><i class="fas fa-star"></i></label>
+            <input type="radio" id="star3" name="grade" value="5">
+            <label for="star3"><i class="fas fa-star"></i></label>
+            <input type="radio" id="star2" name="grade" value="5">
+            <label for="star2"><i class="fas fa-star"></i></label>
+            <input type="radio" id="star1" name="grade" value="5">
+            <label for="star1"><i class="fas fa-star"></i></label>
+        </div><br><br>
+
+        <button type="submit">Submit</button>
+    </form>
+
+    <h2>Reviews</h2>
+    <ul id="reviewList">
+        <!-- Reviews will be displayed here -->
+    </ul>
+
+    <script src="https://kit.fontawesome.com/a076d05399.js"></script>
+    <script>
+        document.getElementById('addReviewForm').addEventListener('submit', function(event) {
+            event.preventDefault();
+            var customerId = document.getElementById('customerId').value;
+            var storeId = document.getElementById('storeId').value;
+            var reviewText = document.getElementById('reviewText').value;
+            var grade = document.querySelector('.stars input:checked').value;
+
+            fetch('/reviews/' + customerId + '/' + storeId, {
+                method: 'POST',
+                headers: {
+                    'Content-Type': 'application/json'
+                },
+                body: JSON.stringify({
+                    reviewText: reviewText,
+                    grade: grade
+                })
+            })
+                .then(response => response.json())
+                .then(data => {
+                    console.log('Review added:', data);
+                    // Refresh review list after adding a new review
+                    getReviews();
+                })
+                .catch(error => {
+                    console.error('Error adding review:', error);
+                });
+        });
+
+        function getReviews() {
+            fetch('/reviews')
+                .then(response => response.json())
+                .then(data => {
+                    console.log('Reviews:', data);
+                    // Display reviews
+                    var reviewList = document.getElementById('reviewList');
+                    reviewList.innerHTML = ''; // Clear existing list
+                    data.forEach(review => {
+                        var listItem = document.createElement('li');
+                        listItem.textContent = review.reviewText + ' - Grade: ' + review.grade;
+                        reviewList.appendChild(listItem);
+                    });
+                })
+                .catch(error => {
+                    console.error('Error fetching reviews:', error);
+                });
+        }
+
+        // Initial call to fetch and display reviews
+        getReviews();
+    </script>
+</body>
+</html>
