+++ src/test/java/org/crm/crmproject/Repository/review/CusReviewTests.java
@@ -6,7 +6,6 @@ import org.crm.crmproject.repository.CusReviewRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.security.crypto.password.PasswordEncoder;
 
 import java.util.stream.IntStream;
